pyproject.toml
[project]
name = "easy-kicad"
version = "0.1.0"
description = "Desktop GUI for importing LCSC parts into KiCad with preview support"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "easyeda2kicad>=0.8.0",
    "fastapi>=0.128.8",
    "platformdirs>=4.4.0",
    "pydantic>=2.12.5",
    "pywebview>=6.1",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.39.0",
]

[project.scripts]
easy-kicad = "easy_kicad:main"

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
]
